requirements.txt
torch
torchvision
langchain
langchain-groq
langchain-google-genai
langchain-community
langgraph
streamlit
pillow
numpy
python-dotenv
matplotlib
opencv-python
pandas
sentence-transformers
faiss-cpu
tabulate
scikit-learn